🔨 refactor: Rename `my_snak_bar.dart` to `my_snack_bar.dart`
The changes in this commit focus on renaming the `my_snak_bar.dart` file to `my_snack_bar.dart` across multiple views and widgets. This is to ensure consistent naming conventions throughout the codebase, as the correct spelling is "snack bar" rather than "snak bar".

The files affected by this change are:

- `lib/presentation/views/schools/Widgets/add_grade_to_school.dart`
- `lib/presentation/views/class_rooms/widgets/class_rooms_widget.dart`
- `lib/presentation/views/cohort_settings/add_subjects_to_cohort.dart`
- `lib/presentation/views/proctor/widgets/add_new_proctor.dart`
- `lib/presentation/views/privileges/widgets/front_id_screen.dart`
- `lib/presentation/views/control_mission/distribution_and_details/distribution.dart`
- `lib/presentation/views/control_mission/widgets/create_mission_widget.dart`
- `lib/presentation/views/schools/Widgets/add_new_school.dart`

This change helps to improve the codebase's readability and maintainability by ensuring consistent naming conventions.
