+++ lib/domain/controllers/batch_documents.dart/create_covers_sheets_controller.dart
@@ -142,7 +142,7 @@ class CreateCoversSheetsController extends GetxController {
     ResponseHandler<EducationsYearsModel> responseHandler = ResponseHandler();
     Either<Failure, EducationsYearsModel> response =
         await responseHandler.getResponse(
-      path: EducationYearsLinks.educationyear,
+      path: EducationYearsLinks.educationYear,
       converter: EducationsYearsModel.fromJson,
       type: ReqTypeEnum.GET,
     );
