+++ lib/domain/controllers/batch_documents.dart/seat_number_controller.dart
@@ -141,7 +141,7 @@ class SeatNumberController extends GetxController {
     ResponseHandler<EducationsYearsModel> responseHandler = ResponseHandler();
     Either<Failure, EducationsYearsModel> response =
         await responseHandler.getResponse(
-      path: EducationYearsLinks.educationyear,
+      path: EducationYearsLinks.educationYear,
       converter: EducationsYearsModel.fromJson,
       type: ReqTypeEnum.GET,
     );
