+++ lib/domain/controllers/control_mission/control_mission_controller.dart
@@ -67,7 +67,7 @@ class ControlMissionController extends GetxController {
     update();
 
     final response = await ResponseHandler<EducationsYearsModel>().getResponse(
-      path: EducationYearsLinks.educationyear,
+      path: EducationYearsLinks.educationYear,
       converter: EducationsYearsModel.fromJson,
       type: ReqTypeEnum.GET,
     );
