+++ lib/domain/controllers/control_mission/create_control_mission.dart
@@ -228,7 +228,7 @@ class CreateControlMissionController extends GetxController {
     update();
 
     final response = await ResponseHandler<EducationsYearsModel>().getResponse(
-      path: EducationYearsLinks.educationyear,
+      path: EducationYearsLinks.educationYear,
       converter: EducationsYearsModel.fromJson,
       type: ReqTypeEnum.GET,
     );
