+++ lib/presentation/views/Login/widgets/select_school_form.dart
@@ -8,7 +8,7 @@ import '../../../../Data/Models/school/school_type/school_type_model.dart';
 import '../../../../domain/controllers/profile_controller.dart';
 import '../../../../domain/controllers/school_controller.dart';
 import '../../../resource_manager/ReusableWidget/loading_indicators.dart';
-import '../../../resource_manager/ReusableWidget/my_snak_bar.dart';
+import '../../../resource_manager/ReusableWidget/my_snack_bar.dart';
 import '../../../resource_manager/routes/app_routes_names_and_paths.dart';
 
 class SelectSchoolForm extends GetView<ProfileController> {
