+++ lib/presentation/views/admin_screen/widgets/admin_widget.dart
@@ -10,7 +10,7 @@ import '../../../../domain/controllers/admin_controller.dart';
 import '../../../../domain/controllers/profile_controller.dart';
 import '../../../resource_manager/ReusableWidget/app_dialogs.dart';
 import '../../../resource_manager/ReusableWidget/loading_indicators.dart';
-import '../../../resource_manager/ReusableWidget/my_snak_bar.dart';
+import '../../../resource_manager/ReusableWidget/my_snack_bar.dart';
 import '../../../resource_manager/ReusableWidget/show_dialogue.dart';
 import 'add_role_user.dart';
 import 'edit_user_widget.dart';
