+++ lib/presentation/views/batch_documents/widgets/add_new_cover_widget.dart
@@ -9,7 +9,7 @@ import '../../../../domain/controllers/batch_documents.dart/cover_sheets_control
 import '../../../../domain/controllers/batch_documents.dart/create_covers_sheets_controller.dart';
 import '../../../resource_manager/ReusableWidget/drop_down_button.dart';
 import '../../../resource_manager/ReusableWidget/loading_indicators.dart';
-import '../../../resource_manager/ReusableWidget/my_snak_bar.dart';
+import '../../../resource_manager/ReusableWidget/my_snack_bar.dart';
 import '../../../resource_manager/ReusableWidget/show_dialogue.dart';
 import '../../../resource_manager/validations.dart';
 
@@ -488,7 +488,7 @@ class AddNewCoverWidget extends GetView<CreateCoversSheetsController> {
                           controlMissionId:
                               controller.selectedItemControlMission!.value,
                           gradeId: controller.selectedItemGrade!.value,
-                          educationyearId:
+                          educationYearId:
                               controller.selectedItemEducationYear!.value,
                           year: selectedYear!,
                           month: '${selectedDay!} ${selectedMonth!}',
