+++ lib/presentation/views/batch_documents/widgets/cover_widget.dart
@@ -10,7 +10,7 @@ import '../../../../domain/controllers/batch_documents.dart/cover_sheets_control
 import '../../../../domain/controllers/profile_controller.dart';
 import '../../../resource_manager/ReusableWidget/app_dialogs.dart';
 import '../../../resource_manager/ReusableWidget/loading_indicators.dart';
-import '../../../resource_manager/ReusableWidget/my_snak_bar.dart';
+import '../../../resource_manager/ReusableWidget/my_snack_bar.dart';
 import '../../../resource_manager/ReusableWidget/show_dialogue.dart';
 import 'edit_cover__ad_widget.dart';
 import 'edit_cover_widget.dart';
