+++ lib/presentation/views/batch_documents/widgets/edit_cover__ad_widget.dart
@@ -10,7 +10,7 @@ import '../../../../app/extensions/convert_date_string_to_iso8601_string_extensi
 import '../../../../domain/controllers/batch_documents.dart/edit_cover_controller.dart';
 import '../../../resource_manager/ReusableWidget/drop_down_button.dart';
 import '../../../resource_manager/ReusableWidget/loading_indicators.dart';
-import '../../../resource_manager/ReusableWidget/my_snak_bar.dart';
+import '../../../resource_manager/ReusableWidget/my_snack_bar.dart';
 import '../../../resource_manager/validations.dart';
 
 // ignore: must_be_immutable
