+++ lib/presentation/views/batch_documents/widgets/edit_cover_widget.dart
@@ -9,7 +9,7 @@ import '../../../../Data/Models/exam_mission/exam_mission_res_model.dart';
 import '../../../../domain/controllers/batch_documents.dart/edit_cover_controller.dart';
 import '../../../resource_manager/ReusableWidget/drop_down_button.dart';
 import '../../../resource_manager/ReusableWidget/loading_indicators.dart';
-import '../../../resource_manager/ReusableWidget/my_snak_bar.dart';
+import '../../../resource_manager/ReusableWidget/my_snack_bar.dart';
 import '../../../resource_manager/validations.dart';
 
 // ignore: must_be_immutable
