+++ lib/presentation/views/class_rooms/widgets/class_rooms_widget.dart
@@ -8,7 +8,7 @@ import '../../../../domain/controllers/class_room_controller.dart';
 import '../../../../domain/controllers/profile_controller.dart';
 import '../../../resource_manager/ReusableWidget/app_dialogs.dart';
 import '../../../resource_manager/ReusableWidget/loading_indicators.dart';
-import '../../../resource_manager/ReusableWidget/my_snak_bar.dart';
+import '../../../resource_manager/ReusableWidget/my_snack_bar.dart';
 import '../../../resource_manager/ReusableWidget/show_dialogue.dart';
 import 'edit_class_room_widget.dart';
