+++ lib/presentation/views/class_rooms/widgets/edit_class_room_widget.dart
@@ -7,7 +7,7 @@ import '../../../../domain/controllers/class_room_controller.dart';
 import '../../../resource_manager/ReusableWidget/elevated_back_button.dart';
 import '../../../resource_manager/ReusableWidget/elevated_edit_button.dart';
 import '../../../resource_manager/ReusableWidget/loading_indicators.dart';
-import '../../../resource_manager/ReusableWidget/my_snak_bar.dart';
+import '../../../resource_manager/ReusableWidget/my_snack_bar.dart';
 import '../../../resource_manager/ReusableWidget/my_text_form_field.dart';
 import '../../../resource_manager/validations.dart';
 import '../../class_room_seats/widgets/render_seat_widget.dart';
