+++ lib/presentation/views/cohort_settings/add_subjects_to_cohort.dart
@@ -10,7 +10,7 @@ import '../../../domain/controllers/operation_cohort_controller.dart';
 import '../../../domain/controllers/subject/subject_controller.dart';
 import '../../resource_manager/ReusableWidget/drop_down_button.dart';
 import '../../resource_manager/ReusableWidget/loading_indicators.dart';
-import '../../resource_manager/ReusableWidget/my_snak_bar.dart';
+import '../../resource_manager/ReusableWidget/my_snack_bar.dart';
 import '../../resource_manager/ReusableWidget/show_dialogue.dart';
 
 class AddSubjectsToCohort extends GetView<CohortsSettingsController> {
