+++ lib/presentation/views/cohort_settings/widgets/edit_cohort_widget.dart
@@ -10,7 +10,7 @@ import 'package:multi_dropdown/models/value_item.dart';
 import '../../../resource_manager/ReusableWidget/drop_down_button.dart';
 import '../../../resource_manager/ReusableWidget/elevated_back_button.dart';
 import '../../../resource_manager/ReusableWidget/loading_indicators.dart';
-import '../../../resource_manager/ReusableWidget/my_snak_bar.dart';
+import '../../../resource_manager/ReusableWidget/my_snack_bar.dart';
 
 class EditCohortWidget extends GetView<OperationCohortController> {
   final CohortResModel cohort;
