+++ lib/presentation/views/cohort_settings/widgets/operation_cohort_screen.dart
@@ -9,7 +9,7 @@ import '../../../../domain/controllers/operation_cohort_controller.dart';
 import '../../../../domain/controllers/profile_controller.dart';
 import '../../../resource_manager/ReusableWidget/app_dialogs.dart';
 import '../../../resource_manager/ReusableWidget/loading_indicators.dart';
-import '../../../resource_manager/ReusableWidget/my_snak_bar.dart';
+import '../../../resource_manager/ReusableWidget/my_snack_bar.dart';
 import '../../../resource_manager/ReusableWidget/show_dialogue.dart';
 import '../add_subjects_to_cohort.dart';
 import 'add_cohort_widget.dart';
