+++ lib/presentation/views/control_mission/distribution_and_details/distribution.dart
@@ -12,7 +12,7 @@ import '../../../../domain/controllers/profile_controller.dart';
 import '../../../resource_manager/ReusableWidget/app_dialogs.dart';
 import '../../../resource_manager/ReusableWidget/header_widget.dart';
 import '../../../resource_manager/ReusableWidget/loading_indicators.dart';
-import '../../../resource_manager/ReusableWidget/my_snak_bar.dart';
+import '../../../resource_manager/ReusableWidget/my_snack_bar.dart';
 import '../../../resource_manager/ReusableWidget/show_dialogue.dart';
 import '../../../resource_manager/routes/app_routes_names_and_paths.dart';
 import '../widgets/add_exam_room_widget.dart';
