+++ lib/presentation/views/privileges/widgets/front_id_screen.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
 import '../../../../domain/controllers/privileges_controller.dart';
-import '../../../resource_manager/ReusableWidget/my_snak_bar.dart';
+import '../../../resource_manager/ReusableWidget/my_snack_bar.dart';
 
 class FrontIdScreen extends GetView<PrivilegesController> {
   final Color color;
