+++ lib/presentation/views/schools/Widgets/add_grade_to_school.dart
@@ -7,7 +7,7 @@ import '../../../../domain/controllers/school_controller.dart';
 import '../../../resource_manager/ReusableWidget/elevated_add_button.dart';
 import '../../../resource_manager/ReusableWidget/elevated_back_button.dart';
 import '../../../resource_manager/ReusableWidget/loading_indicators.dart';
-import '../../../resource_manager/ReusableWidget/my_snak_bar.dart';
+import '../../../resource_manager/ReusableWidget/my_snack_bar.dart';
 
 class AddNewGradeToSchool extends GetView<SchoolController> {
   const AddNewGradeToSchool({
