+++ lib/presentation/views/schools/Widgets/add_new_school.dart
@@ -8,7 +8,7 @@ import '../../../resource_manager/ReusableWidget/drop_down_button.dart';
 import '../../../resource_manager/ReusableWidget/elevated_add_button.dart';
 import '../../../resource_manager/ReusableWidget/elevated_back_button.dart';
 import '../../../resource_manager/ReusableWidget/loading_indicators.dart';
-import '../../../resource_manager/ReusableWidget/my_snak_bar.dart';
+import '../../../resource_manager/ReusableWidget/my_snack_bar.dart';
 
 class AddNewSchoolWidget extends GetView<SchoolController> {
   const AddNewSchoolWidget({
