+++ lib/presentation/views/student/widgets/add_single_student_widget.dart
@@ -7,7 +7,7 @@ import 'package:multi_dropdown/models/value_item.dart';
 import '../../../../domain/controllers/students_controllers/add_new_student_controller.dart';
 import '../../../resource_manager/ReusableWidget/drop_down_button.dart';
 import '../../../resource_manager/ReusableWidget/loading_indicators.dart';
-import '../../../resource_manager/ReusableWidget/my_snak_bar.dart';
+import '../../../resource_manager/ReusableWidget/my_snack_bar.dart';
 import '../../../resource_manager/ReusableWidget/my_text_form_field.dart';
 import '../../../resource_manager/validations.dart';
