+++ lib/presentation/views/subject_setting/widgets/edit_operation_widget.dart
@@ -9,7 +9,7 @@ import '../../../../domain/controllers/subject/operation_controller.dart';
 import '../../../../domain/controllers/subject/subject_controller.dart';
 import '../../../resource_manager/ReusableWidget/drop_down_button.dart';
 import '../../../resource_manager/ReusableWidget/loading_indicators.dart';
-import '../../../resource_manager/ReusableWidget/my_snak_bar.dart';
+import '../../../resource_manager/ReusableWidget/my_snack_bar.dart';
 
 class EditOperationWidget extends GetView<EditSubjectsController> {
   final SubjectResModel subjectResModel;
