+++ lib/presentation/views/subject_setting/widgets/operation_widget.dart
@@ -9,7 +9,7 @@ import '../../../../domain/controllers/controllers.dart';
 import '../../../resource_manager/ReusableWidget/app_dialogs.dart';
 import '../../../resource_manager/ReusableWidget/loading_indicators.dart';
 import '../../../resource_manager/ReusableWidget/my_back_button.dart';
-import '../../../resource_manager/ReusableWidget/my_snak_bar.dart';
+import '../../../resource_manager/ReusableWidget/my_snack_bar.dart';
 import '../../../resource_manager/ReusableWidget/show_dialogue.dart';
 import 'edit_operation_widget.dart';
